[W1 수정중] - 가독성 좋은 코드로 변경 - 상속 관계 매핑에 따른 코드 변경

+++ src/main/java/com/pet/comes/dto/Req/ScheduleDto.java
@@ -28,7 +28,7 @@ public class ScheduleDto {
 
 	private String disease;
 	private long dogId;
-	private float weight;
+	private double weight;
 	private String address;
 	private String x;
 	private String y;
